making matches and updating user stats working nicely, also formated the date in the viewDB controller, this formatter util is in matchUtils, working on all the view/getters for matches
