+++ package.json
@@ -30,6 +30,7 @@
     "@prisma/client": "^6.5.0",
     "better-sqlite3": "^11.9.0",
     "close-with-grace": "^2.2.0",
+    "dayjs": "^1.11.13",
     "fastify": "^5.2.1",
     "fastify-file-upload": "^4.0.0",
     "form-data": "^4.0.2",
