+++ server/controllers/dev/userDev/matchDev.controller.ts
@@ -1,16 +1,6 @@
 import { FastifyRequest, FastifyReply } from 'fastify';
 import { updateWins, updateLosses } from "../../user/stats.controller";
 
-// add matches
-// check updates of user stats
-// view matchs of user
-// view matchs of users' friends / specific other player
-
-// figure date stuff out
-//const date = new Date(request.server.prisma.match.date);
-//const formatted = date.toLocaleDateString('en-GB');
-
-
 export const makeMatch = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
 	try {
 		const { won, lost } = request.params as { won: number, lost: number };
@@ -24,10 +14,10 @@ export const makeMatch = async (request: FastifyRequest, reply: FastifyReply): P
 				date: new Date(),
 			},
 		});
-
+		
 		await updateWins(wonId, request, reply);
 		await updateLosses(lostId, request, reply);
-
+		
 		return reply.code(200).send({ message: "Successfully saved played match!" });
 	} catch (error) {
 		reply.status(500).send({ error: 'Failed to save match in database' });
@@ -35,12 +25,48 @@ export const makeMatch = async (request: FastifyRequest, reply: FastifyReply): P
 };
 
 
-export const viewMatches = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+// viewUserMatch
+export const viewUserMatch = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
 	try {
-	  const matches = await request.server.prisma.match.findMany();
 
-	  reply.send({ matches: matches });
-	} catch (error) {
-	  reply.status(500).send({ error: 'Failed to fetch match data' });
+	} catch {
+		reply.status(500).send({ error: 'Failed to view match history' });
 	}
-  };
+};
+
+// viewWonMatch
+export const viewWonMatch = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+
+	} catch {
+		reply.status(500).send({ error: 'Failed to view match history' });
+	}
+};
+
+// viewLostMatch
+export const viewLostMatch = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+
+	} catch {
+		reply.status(500).send({ error: 'Failed to view match history' });
+	}
+};
+
+// viewFriendMatch
+export const viewFriendMatch = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+
+	} catch {
+		reply.status(500).send({ error: 'Failed to view match history' });
+	}
+};
+
+// viewFriendvsUser
+export const viewFvsU = async (request: FastifyRequest, reply: FastifyReply): Promise<any> => {
+	try {
+
+	} catch {
+		reply.status(500).send({ error: 'Failed to view match history' });
+	}
+};
+
