+++ server/controllers/utils/friendUtils.controller.ts
@@ -104,6 +104,34 @@ export async function getBlocked(userId: number, request: FastifyRequest) {
 	return blockedPlayers;
 };
 
+export async function getAllFriends(userId: number, request: FastifyRequest) {
+	const friendships = await request.server.prisma.friend.findMany({
+		where: {
+			status: { in: ['FRIENDS', 'PENDING', 'BLOCKED'] },
+			OR: [
+				{ user1Id: userId},
+				{ user2Id: userId},
+			],
+		},
+		include: {
+			user1: true, 
+			user2: true,
+		},
+	});
+	const friends = friendships.map(f => {
+		const friendUser = f.user1Id === userId ? f.user2 : f.user1;
+		return {
+			friend: {
+				username: friendUser.username,
+				status: friendUser.status,
+			  },
+			status: f.status,
+			initiator: f.user1Id,
+		};
+	});
+	return friends;
+};
+
 
 
 // sarah notes - ill get back to this
