+++ server/routes/routes.ts
@@ -4,7 +4,7 @@ import { FastifyInstance, FastifyRequest, FastifyReply } from 'fastify';
 	import {getHome} from "../controllers/dev/getHome.controller";
 	import {populate, viewDB, viewID} from "../controllers/dev/userDev/view.controller";
 	import {sendReqDev, acceptReqDev, rejectReqDev} from "../controllers/dev/userDev/friendRequests.controller";
-	import {viewMatches, makeMatch} from "../controllers/dev/userDev/matchDev.controller";
+	import {makeMatch, viewUserMatch, viewWonMatch, viewLostMatch, viewFriendMatch, viewFvsU} from "../controllers/dev/userDev/matchDev.controller";
 
 	// web
 	import {dash} from "../controllers/web/dash.controller";
@@ -15,6 +15,7 @@ import { FastifyInstance, FastifyRequest, FastifyReply } from 'fastify';
 	import {editAvatar, deleteAvatar} from "../controllers/user/avatar.controller";
 	import {sendFriendReq, acceptFriendReq, rejectFriendReq, blockFriend, viewPlayers} from "../controllers/user/friends.controller";
 	import {saveMatch} from "../controllers/user/stats.controller";
+	// , viewUserMatches, viewWonMatches, viewLostMatches, viewFriendMatches, viewFriendvsUser
 
 // game endpoints
 import { getPlayer, getPlayerInfo, updatePlayer, syncPlayers } from "../controllers/game/player.controller";
@@ -27,8 +28,12 @@ async function routes(fastify: FastifyInstance) {
   fastify.get('/viewDB', viewDB);
   fastify.get('/viewID', viewID);
   
-  fastify.get('/viewMatches', viewMatches);
   fastify.get('/makeMatch/:won/:lost', makeMatch);
+  fastify.get('/viewUserMatch', viewUserMatch);
+  fastify.get('/viewWonMatch', viewWonMatch);
+  fastify.get('/viewLostMatch', viewLostMatch);
+  fastify.get('/viewFriendMatch/:firendId', viewFriendMatch);
+  fastify.get('/viewFvsU/:firendId', viewFvsU);
   
   fastify.get('/sendReqD/:receiverId/:userId', sendReqDev);
   fastify.get('/acceptReqD/:senderId/:userId', acceptReqDev);
@@ -56,6 +61,11 @@ async function routes(fastify: FastifyInstance) {
   fastify.post('/deleteAvatar/:username', deleteAvatar);
   // user statistics/matches endpoints
   fastify.post('/saveMatch/:won/:lost', saveMatch);
+//  fastify.get('/viewUserMatches', viewUserMatches);
+//  fastify.get('/viewWonMatches', viewWonMatches);
+//  fastify.get('/viewLostMatches', viewLostMatches);
+//  fastify.get('/viewFriendMatches/:friendId', viewFriendMatches);
+//  fastify.get('/viewFriendvsUser/:friendId', viewFriendvsUser);
 
   // game endpoints
   // online users, game history, game options...
